Separate TransactionAuthenticator trait implementation from the solana tx parsing/authentication implementation utils
